update focused_geometry_editor Now it's just create controls, pass geometry to and from draw tools
