--- src/features/focused_geometry_editor/atoms/isEditorActive.ts
@@ -1,3 +1,0 @@
-import { createBooleanAtom } from '~utils/atoms/createPrimitives';
-
-export const isEditorActiveAtom = createBooleanAtom(false, 'isEditorActiveAtom');
