+++ src/features/focused_geometry_editor/constants.ts
@@ -1,0 +1,2 @@
+export const FOCUSED_GEOMETRY_EDITOR_CONTROL_ID = 'FreehandGeometry';
+export const FOCUSED_GEOMETRY_EDITOR_CONTROL_NAME = 'Freehand Geometry';
